refactor(windy-plugin): drop svelte
