+++ libs/windy-sounding/package.json
@@ -35,13 +35,11 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "catalog:default",
-    "@sveltejs/vite-plugin-svelte": "^7.2.0",
     "date-fns": "catalog:default",
     "geojson": "^0.5.0",
     "preact": "^10.29.7",
     "react-redux": "^9.3.0",
-    "jsonc-eslint-parser": "catalog:default",
-    "svelte": "^5.56.7"
+    "jsonc-eslint-parser": "catalog:default"
   },
   "devDependencies": {
     "@babel/core": "^7.29.7",
@@ -49,7 +47,6 @@
     "@babel/preset-typescript": "^7.29.7",
     "@preact/preset-vite": "catalog:",
     "@types/geojson": "^7946.0.15",
-    "svelte-check": "^4.7.3",
     "vite": "catalog:default"
   }
 }
